+++ package/lib/src/helpers/enabled_controls.dart
@@ -1,0 +1,11 @@
+/// this class helps you to hide some player buttons
+class EnabledControls {
+  final bool seekSwipes, volumeSwipes,doubleTapToSeek,desktopDoubleTapToFullScreen;
+
+  const EnabledControls({
+    this.seekSwipes = true,
+    this.volumeSwipes=true,
+    this.doubleTapToSeek=true,
+    this.desktopDoubleTapToFullScreen=true
+  });
+}